PyDa/discord_code/requirements.txt
wolframalpha
wikipedia
SpeechRecognition
discord.py
python-dotenv
PyDrive2